requirements.txt
fastexcel==0.12.1
genai==2.1.0
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
polars==1.22.0
pydantic==2.10.6
pypdf==5.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
streamlit==1.42.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
XlsxWriter==3.2.2
yarl==1.18.3
zipp==1.0.0
